席の解除のフック化
