+++ booking_web/src/components/atoms/button/RemoveTableRelationButton.jsx
@@ -1,49 +1,16 @@
 import React, { useEffect, useState } from "react";
-import { useSetRecoilState } from "recoil";
 import { Box, Button, Select } from "@chakra-ui/react";
 
-import { today } from "../../../utils/date";
-import { removeTableRelation, getAllBooking } from "../../../utils/api";
-import {
-  allBookingState,
-  todayBookingState,
-  noAssignedBookingState,
-} from "../../../atom/state";
-import { useMessage } from "../../../hooks/useMessage";
+import { useRemoveTableRelation } from "../../../hooks/useRemoveTableRelation";
 
 export const RemoveTableRelationButton = (props) => {
   const { bookingId, tableNum } = props;
-  const setAllBooking = useSetRecoilState(allBookingState);
-  const setTodayBooking = useSetRecoilState(todayBookingState);
-  const setNoAssigendBooking = useSetRecoilState(noAssignedBookingState);
-  const { showMessage } = useMessage();
-
-
   const [table, setTable] = useState();
+  const { removeTable } = useRemoveTableRelation();
 
   const onClickRemoveTableRelation = async () => {
-    const data = {
-      booking: {
-        id: bookingId,
-      },
-      table: {
-        id: table,
-      },
-    };
-    console.log(table);
-    await removeTableRelation(data);
-    const allBooking = await getAllBooking();
-    setAllBooking(allBooking);
-    const todayBooking = allBooking.filter(
-      (booking) => booking.date === today()
-    );
-    setTodayBooking(todayBooking);
-    const noAssigendBooking = todayBooking.filter((booking) => {
-      return booking.tables.length === 0;
-    });
-    setNoAssigendBooking(noAssigendBooking);
-    showMessage({ title: "席を解除しました", status: "success" });
-  };
+    removeTable(bookingId, table);
+  }
 
   useEffect(() => {
     if (tableNum.length > 0) {
